Added screen capturing as fallback to videos, for pre-KitKat devices.

+++ src/main/java/com/shazam/fork/runtime/ScreenRecorderStopper.java
@@ -12,27 +12,40 @@ package com.shazam.fork.runtime;
 
 import com.android.ddmlib.IDevice;
 import com.android.ddmlib.NullOutputReceiver;
+import com.shazam.fork.system.CancellableShellOutputReceiver;
 import com.shazam.fork.system.CollectingShellOutputReceiver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-class ScreenRecordStopper {
-    private static final Logger logger = LoggerFactory.getLogger(ScreenRecordStopper.class);
+import javax.annotation.Nullable;
+
+import static org.apache.commons.lang3.StringUtils.isBlank;
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
+
+class ScreenRecorderStopper {
+    private static final Logger logger = LoggerFactory.getLogger(ScreenRecorderStopper.class);
     private final NullOutputReceiver nullOutputReceiver = new NullOutputReceiver();
     private final IDevice deviceInterface;
+    private final CancellableShellOutputReceiver cancellableReceiver;
 
-    ScreenRecordStopper(IDevice deviceInterface) {
+    ScreenRecorderStopper(IDevice deviceInterface, CancellableShellOutputReceiver cancellableReceiver) {
         this.deviceInterface = deviceInterface;
+        this.cancellableReceiver = cancellableReceiver;
     }
 
-    void interruptScreenRecord() {
+    void cancelScreenRecord() {
+        cancellableReceiver.cancel();
+    }
+
+    void stopScreenRecord() {
         CollectingShellOutputReceiver receiver = new CollectingShellOutputReceiver();
         try {
             deviceInterface.executeShellCommand("ps |grep screenrecord", receiver);
             String pid = extractPidOfScreenrecordProcess(receiver);
-            deviceInterface.executeShellCommand("kill -2 " + pid, nullOutputReceiver);
-
+            if (isNotBlank(pid)) {
+                deviceInterface.executeShellCommand("kill -2 " + pid, nullOutputReceiver);
+            }
         } catch (Exception e) {
             logger.error("Error while killing recording processes", e);
         }
@@ -40,14 +53,20 @@ class ScreenRecordStopper {
 
     /**
      * The output of ps is of the format:
-     * <br><code>USER PID PPID VSIZE RSS WCHAN PC NAME</code></br>
+     * <br><code>USER PID PPID VSIZE RSS WCHAN PC NAME</code>
+     * <br>
      *
-     * For now, we're going to assume there's only one of them at a time.
+     * <b>NOTE: For now, we're going to assume there's only one screenrecord process at a time.</b>
+     *
      * @param receiver the command receiver
      * @return the string representing the screenrecord process PID
      */
+    @Nullable
     private String extractPidOfScreenrecordProcess(CollectingShellOutputReceiver receiver) {
         String output = receiver.getOutput();
+        if (isBlank(output)) {
+            return null;
+        }
         String[] split = output.split("\\s+");
         String pid = split[1];
         logger.trace("Extracted PID {} from output {}", pid, output);
